Move legacy view code into new file
